Adding updating file

+++ 16th Oct 2023/Task3Triangle Classifier.py
@@ -7,14 +7,16 @@ side 1, side 2 and side 3
 output - Eq, Iso, Scalene -
 Eq. = side 1 == side 2 = side 3"""
 
-side1 = int(input("Enter one side1:-"))
-side2 = int(input("Enter one side2:-"))
-side3 = int(input("Enter one side3:-"))
+side1 = float(input("Enter one side1:-"))
+side2 = float(input("Enter one side2:-"))
+side3 = float(input("Enter one side3:-"))
 
 if side1 == side2 == side3:
     print("This is a equaleteral triangle")
 elif side1 != side2 != side3:
     print("This is a scalene triangle")
+    #elif side1 = side2 or side2 = side3 or side1= side3
+    #print("This isa Isosceles triangle")
 else:
     print("This isa Isosceles triangle")
